requirements.txt
aniso8601==9.0.1
apispec==6.3.0
click==8.1.3
Flask==2.2.3
Flask-JWT-Extended==4.4.4
Flask-RESTful==0.3.9
flask-smorest==0.40.0
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.0
passlib==1.7.4
PyJWT==2.6.0
python-dotenv==1.0.0
pytz==2022.7.1
six==1.16.0
SQLAlchemy==2.0.7
sqlitemodel==0.1.3
typing_extensions==4.5.0
webargs==8.2.0
Werkzeug==2.2.3
